+++ packages/core-js/actual/array-buffer/transfer.js
@@ -1,0 +1,2 @@
+require('../../stable/array-buffer');
+require('../../modules/esnext.array-buffer.transfer');