+++ packages/core-js/full/array-buffer/detached.js
@@ -1,0 +1,3 @@
+var parent = require('../../actual/array-buffer/detached');
+
+module.exports = parent;