+++ packages/core-js/full/array-buffer/transfer.js
@@ -1,0 +1,3 @@
+var parent = require('../../actual/array-buffer/transfer');
+
+module.exports = parent;