+++ packages/core-js/proposals/array-buffer-transfer.js
@@ -1,0 +1,3 @@
+require('../modules/esnext.array-buffer.detached');
+require('../modules/esnext.array-buffer.transfer');
+require('../modules/esnext.array-buffer.transfer-to-fixed-length');